Parameterize the encrypted file path

+++ rakefile.rb
@@ -1,7 +1,23 @@
 task :default => [:help]
 
+def settings_file_path
+    current_dir = File.expand_path(File.dirname(__FILE__))
+    settings_path = File.join(current_dir, './.settings')
+    File.expand_path(settings_path)
+end
+
 def encrypted_file
-    return File.expand_path "~/Documents/t/t.zip"
+    unless File.exists? settings_file_path
+        raise "Settings file not found at '#{settings_file_path}'. Make sure to run rake setup"
+    end
+
+    File.read(settings_file_path).strip
+end
+
+def encrypted_file_configured?
+    return false unless File.exists?(settings_file_path)
+
+    return !File.read(settings_file_path).strip.empty?
 end
 
 def encrypted_file_bck
@@ -38,10 +54,16 @@ task :help do
     puts ""
     puts "Settings"
     puts "="*16
-    puts "encrypted_file               => #{encrypted_file}"
-    puts "encrypted_file_parent        => #{encrypted_file_parent}"
-    puts "decrypted_folder_parent      => #{decrypted_folder_parent}"
-    puts "decrypted_folder_full_path   => #{decrypted_folder_full_path}"
+    puts "settings_file_path           => #{settings_file_path}"
+
+    if !encrypted_file_configured?
+        puts "Settings not configured yet. Run rake setup"
+    else
+        puts "encrypted_file               => #{encrypted_file}"
+        puts "encrypted_file_parent        => #{encrypted_file_parent}"
+        puts "decrypted_folder_parent      => #{decrypted_folder_parent}"
+        puts "decrypted_folder_full_path   => #{decrypted_folder_full_path}"
+    end
 end
 
 task :e => :encrypt
@@ -82,15 +104,28 @@ task :c => :clean
 task :clean do
     rm_rf decrypted_folder_full_path
 
-    if File.file? encrypted_file_bck
-        rm encrypted_file_bck
+    if encrypted_file_configured?
+        if File.file? encrypted_file_bck
+            rm encrypted_file_bck
+        end
     end
 end
 
-task :setup do
+task :setup => :configure_path do
     ["clean", "decrypt", "encrypt"].each { |task_name|
         generate_task_alias_wrapper task_name
     }
+end
+
+task :configure_path do
+    puts 'Enter the path of your encryped zip file'
+    file_path_raw = STDIN.gets.strip
+    file_path = File.expand_path file_path_raw
+    unless File.exists? file_path
+        fail "File does not exists at '#{file_path}'"
+    end
+
+    write_file(settings_file_path, file_path)
 end
 
 def generate_task_alias_wrapper(task_name)
